restructured index.js files
