+++ src/views/admin-pages/campaign-manager/index.js
@@ -1,0 +1,1 @@
+export { default as CampaignManager } from './campaign-manager';
