+++ src/views/client-pages/client-manager/index.js
@@ -1,0 +1,1 @@
+export { default as ClientManager } from './client-manager';
