+++ src/views/components/admin-route/index.js
@@ -1,0 +1,1 @@
+export { default as AdminRoute } from './admin-route';
