+++ src/views/components/client-route/index.js
@@ -1,0 +1,1 @@
+export { default as ClientRoute } from './client-route';
