+++ src/views/components/index.js
@@ -1,3 +1,3 @@
-export { default as AdminRoute } from './admin-route/admin-route';
-export { default as ClientRoute } from './client-route/client-route';
-export { default as NotAuthRoute } from './not-auth-route/not-auth-route';
+export { AdminRoute } from './admin-route';
+export { ClientRoute } from './client-route';
+export { NotAuthRoute } from './not-auth-route';
