+++ src/views/components/not-auth-route/index.js
@@ -1,0 +1,1 @@
+export { default as NotAuthRoute } from './not-auth-route';
